third_party/WebKit/LayoutTests/external/wpt/service-workers/cache-storage/serviceworker/cache-keys-attributes-for-service-worker.https-expected.txt

This is a testharness.js-based test.
FAIL Request.IsReloadNavigation should persist. assert_equals: expected "original: false, stored: false" but got "original: undefined, stored: undefined"
Harness: the test ran to completion.
